refactor: moved create users upsert endpoint from UserController (/users) to AuthController (/auth/register)
