+++ backend/account-service/src/main/java/com/ita/challenges/account/infrastructure/adapter/web/in/AuthController.java
@@ -1,18 +1,65 @@
 package com.ita.challenges.account.infrastructure.adapter.web.in;
 
+import com.ita.challenges.account.domain.model.User;
+import com.ita.challenges.account.domain.port.out.UserRepository;
 import com.ita.challenges.account.infrastructure.adapter.web.in.dto.AuthUserDto;
+import com.ita.challenges.account.infrastructure.adapter.web.in.dto.GitHubUserResponse;
+import com.ita.challenges.account.infrastructure.adapter.web.in.dto.UserRequest;
+import com.ita.challenges.account.infrastructure.adapter.web.in.dto.UserResponse;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.security.oauth2.core.user.OAuth2User;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
+import org.springframework.web.client.RestTemplate;
+
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/api/account")
 public class AuthController {
 
-   @GetMapping("/me")
+    private final UserRepository repository;
+    private final RestTemplate restTemplate;
+
+    public AuthController(UserRepository repository, RestTemplate restTemplate) {
+        this.repository = repository;
+        this.restTemplate = restTemplate;
+    }
+
+    @PostMapping("/auth/register")
+    public ResponseEntity<UserResponse> create(@RequestBody UserRequest request) {
+        String url = "https://api.github.com/users/" + request.username();
+
+        GitHubUserResponse gitHubUserResponse = restTemplate.getForObject(url, GitHubUserResponse.class);
+
+        if (gitHubUserResponse == null) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+        }
+
+        Optional<User> userOptional = repository.findByUsername(gitHubUserResponse.id());
+        User user;
+        HttpStatus status;
+
+        if (userOptional.isPresent()) {
+            user = userOptional.get().withRole(request.role());
+            status = HttpStatus.OK;
+        } else {
+            user = new User(gitHubUserResponse.id(), request.role());
+            status = HttpStatus.CREATED;
+        }
+
+        User saved = repository.save(user);
+
+        return ResponseEntity.status(status).body(
+                new UserResponse(
+                        saved.userName(),
+                        saved.userRole()
+                )
+        );
+    }
+
+    @GetMapping("/me")
     public ResponseEntity<AuthUserDto> me(@AuthenticationPrincipal OAuth2User user) {
 
         if (user == null) {
